doctors hate him! try this one new weight loss trick
I fixed the problem with the absolutely absurd amount of lines I was using to do this trick.
The main problem that was proposed by this was that it limited the function to only working with 10 items.
The new solution works with however many items you're willing to make your computer suffer through,
and looks a hell of a lot cleaner.