Fixed up the poor score system, and improved card removing when a match is made.
The score is now a variable relying on the length of the deck of matches of each player, rather than being set on it's own.

Instead of having two var.append() and two var.remove() for every match, matches made in your hand will only use one line, and matches made with the other player will use 3 lines.

The 'Wins' line now displays the cards remaining in the deck.

Added a ui.print() call to the end of the script, to fix a bug with the winning condition print statement.