Fixed issues displaying matches
Not sure if this is working right on linux, but the CMD on windows automatically cuts it off and won't let me scroll up to see the last of my matches when the windows is too small.
However- it no longer crashes and asks before you view them, and before exiting. Yay?