modHelper updated to have more documentation
This means I added a docstring to the module, as well as commenting the reasons for why I coded the only function currently in there like I did.
As well as this, I changed the input variable to 'inp' to be consistent with the changes I'm making to other modules.