Updated merge_alternate()
modFile now follows guideline, as well as this, merge_alternate() is now improved.
Previously there was a weird thing you could come across where you could get a list of lists, or a single list, and it wasn't completely predictable.
Now you can select either of these, and the 'endchar' option is available in both of these modes.
I also removed the check to see if the input list is actually a list, because why the hell not?