Bluffing & New matching systems
Whew, lad, this one's a real juicy update here
Alongside fixing some bugs since I didn't test it on my Arch-top,
I added the ability to BLUFF!
When the Bot player asks is you have a card, you have the ability to say yes or no, if you have the card and say no, you have a 3/5 chance of getting called out!
This adds a bit of strategy to the game.
I also change how the in-hand matching system works (but only for the human player)
When it's your turn, you can select a card you have the match for, and it will match it.
The system will not tell you if you have a match, so you have to figure out on your own!
This does NOT end your turn, so don't worry about that.